Updated the entire state machine framework to include a context
This is a very large update to make it so that every state machine now
requires a context.  The context could very well be null and not used
but it is required to be specified.  By leaving the specification up to
the user, the decision of using a null is then left up to the user and
not the framework.